feat: complete factory with test cases coverage

+++ test/CloakEscrowFactory.Deployment.t.sol
@@ -11,10 +11,7 @@ contract CloakEscrowFactoryDeploymentTest is Test, CloakEscrowFactoryFixtures {
     FactoryFixture public fixture;
 
     event EscrowDeployed(
-        bytes32 indexed merchantId, 
-        address indexed escrowAddress, 
-        uint256 counter,
-        address paymentToken
+        bytes32 indexed merchantId, address indexed escrowAddress, uint256 counter, address paymentToken
     );
     event MerchantRegistered(bytes32 indexed merchantId);
 
@@ -26,11 +23,8 @@ contract CloakEscrowFactoryDeploymentTest is Test, CloakEscrowFactoryFixtures {
     // ============ Single Deployment Tests ============
 
     function test_DeployEscrow_Success() public {
-        address predictedAddress = fixture.factory.predictEscrowAddress(
-            MERCHANT_ID_1,
-            MERCHANT_1,
-            address(fixture.usdcToken)
-        );
+        address predictedAddress =
+            fixture.factory.predictEscrowAddress(MERCHANT_ID_1, MERCHANT_1, address(fixture.usdcToken));
 
         // No prank needed - test contract is owner
         vm.expectEmit(true, true, true, true);
@@ -38,11 +32,7 @@ contract CloakEscrowFactoryDeploymentTest is Test, CloakEscrowFactoryFixtures {
         vm.expectEmit(true, true, true, true);
         emit EscrowDeployed(MERCHANT_ID_1, predictedAddress, 0, address(fixture.usdcToken));
 
-        address deployedAddress = fixture.factory.deployEscrow(
-            MERCHANT_ID_1,
-            MERCHANT_1,
-            address(fixture.usdcToken)
-        );
+        address deployedAddress = fixture.factory.deployEscrow(MERCHANT_ID_1, MERCHANT_1, address(fixture.usdcToken));
 
         // Verify deployment
         assertEq(deployedAddress, predictedAddress);
@@ -56,28 +46,16 @@ contract CloakEscrowFactoryDeploymentTest is Test, CloakEscrowFactoryFixtures {
 
         // Verify escrow configuration
         verifyEscrowConfiguration(
-            deployedAddress,
-            MERCHANT_1,
-            address(fixture.usdcToken),
-            PLATFORM_ADDRESS,
-            DEFAULT_OWNER
+            deployedAddress, MERCHANT_1, address(fixture.usdcToken), PLATFORM_ADDRESS, DEFAULT_OWNER
         );
     }
 
     function test_DeployEscrow_SuccessWithDAI() public {
         // No prank needed - test contract is owner
-        address deployedAddress = fixture.factory.deployEscrow(
-            MERCHANT_ID_1,
-            MERCHANT_1,
-            address(fixture.daiToken)
-        );
+        address deployedAddress = fixture.factory.deployEscrow(MERCHANT_ID_1, MERCHANT_1, address(fixture.daiToken));
 
         verifyEscrowConfiguration(
-            deployedAddress,
-            MERCHANT_1,
-            address(fixture.daiToken),
-            PLATFORM_ADDRESS,
-            DEFAULT_OWNER
+            deployedAddress, MERCHANT_1, address(fixture.daiToken), PLATFORM_ADDRESS, DEFAULT_OWNER
         );
     }
 
@@ -85,18 +63,10 @@ contract CloakEscrowFactoryDeploymentTest is Test, CloakEscrowFactoryFixtures {
         // No prank needed - test contract is owner
 
         // Deploy for merchant 1
-        address escrow1 = fixture.factory.deployEscrow(
-            MERCHANT_ID_1,
-            MERCHANT_1,
-            address(fixture.usdcToken)
-        );
+        address escrow1 = fixture.factory.deployEscrow(MERCHANT_ID_1, MERCHANT_1, address(fixture.usdcToken));
 
         // Deploy for merchant 2
-        address escrow2 = fixture.factory.deployEscrow(
-            MERCHANT_ID_2,
-            MERCHANT_2,
-            address(fixture.usdcToken)
-        );
+        address escrow2 = fixture.factory.deployEscrow(MERCHANT_ID_2, MERCHANT_2, address(fixture.usdcToken));
 
         // No prank used
 
@@ -114,24 +84,16 @@ contract CloakEscrowFactoryDeploymentTest is Test, CloakEscrowFactoryFixtures {
         // First deployment - should emit MerchantRegistered
         vm.expectEmit(true, true, true, true);
         emit MerchantRegistered(MERCHANT_ID_1);
-        address escrow1 = fixture.factory.deployEscrow(
-            MERCHANT_ID_1,
-            MERCHANT_1,
-            address(fixture.usdcToken)
-        );
+        address escrow1 = fixture.factory.deployEscrow(MERCHANT_ID_1, MERCHANT_1, address(fixture.usdcToken));
 
         // Second deployment - should NOT emit MerchantRegistered
         vm.recordLogs();
-        address escrow2 = fixture.factory.deployEscrow(
-            MERCHANT_ID_1,
-            MERCHANT_1,
-            address(fixture.daiToken)
-        );
+        address escrow2 = fixture.factory.deployEscrow(MERCHANT_ID_1, MERCHANT_1, address(fixture.daiToken));
 
         // Verify MerchantRegistered was not emitted for second deployment
         Vm.Log[] memory logs = vm.getRecordedLogs();
         bool merchantRegisteredEmitted = false;
-        for (uint i = 0; i < logs.length; i++) {
+        for (uint256 i = 0; i < logs.length; i++) {
             if (logs[i].topics[0] == keccak256("MerchantRegistered(bytes32)")) {
                 merchantRegisteredEmitted = true;
                 break;
@@ -157,13 +119,9 @@ contract CloakEscrowFactoryDeploymentTest is Test, CloakEscrowFactoryFixtures {
 
         for (uint256 i = 0; i < 5; i++) {
             assertEq(fixture.factory.getMerchantCounter(MERCHANT_ID_1), i);
-            
-            fixture.factory.deployEscrow(
-                MERCHANT_ID_1,
-                MERCHANT_1,
-                address(fixture.usdcToken)
-            );
-            
+
+            fixture.factory.deployEscrow(MERCHANT_ID_1, MERCHANT_1, address(fixture.usdcToken));
+
             assertEq(fixture.factory.getMerchantCounter(MERCHANT_ID_1), i + 1);
             assertEq(fixture.factory.getMerchantEscrowCount(MERCHANT_ID_1), i + 1);
         }
@@ -174,50 +132,26 @@ contract CloakEscrowFactoryDeploymentTest is Test, CloakEscrowFactoryFixtures {
     // ============ Address Prediction Tests ============
 
     function test_PredictEscrowAddress_Accuracy() public {
-        address predicted = fixture.factory.predictEscrowAddress(
-            MERCHANT_ID_1,
-            MERCHANT_1,
-            address(fixture.usdcToken)
-        );
+        address predicted = fixture.factory.predictEscrowAddress(MERCHANT_ID_1, MERCHANT_1, address(fixture.usdcToken));
 
         // No prank needed - test contract is owner
-        address deployed = fixture.factory.deployEscrow(
-            MERCHANT_ID_1,
-            MERCHANT_1,
-            address(fixture.usdcToken)
-        );
+        address deployed = fixture.factory.deployEscrow(MERCHANT_ID_1, MERCHANT_1, address(fixture.usdcToken));
 
         assertEq(predicted, deployed);
     }
 
     function test_PredictEscrowAddress_DifferentMerchants() public {
-        address predicted1 = fixture.factory.predictEscrowAddress(
-            MERCHANT_ID_1,
-            MERCHANT_1,
-            address(fixture.usdcToken)
-        );
+        address predicted1 = fixture.factory.predictEscrowAddress(MERCHANT_ID_1, MERCHANT_1, address(fixture.usdcToken));
 
-        address predicted2 = fixture.factory.predictEscrowAddress(
-            MERCHANT_ID_2,
-            MERCHANT_2,
-            address(fixture.usdcToken)
-        );
+        address predicted2 = fixture.factory.predictEscrowAddress(MERCHANT_ID_2, MERCHANT_2, address(fixture.usdcToken));
 
         assertTrue(predicted1 != predicted2);
     }
 
     function test_PredictEscrowAddress_DifferentTokens() public {
-        address predicted1 = fixture.factory.predictEscrowAddress(
-            MERCHANT_ID_1,
-            MERCHANT_1,
-            address(fixture.usdcToken)
-        );
+        address predicted1 = fixture.factory.predictEscrowAddress(MERCHANT_ID_1, MERCHANT_1, address(fixture.usdcToken));
 
-        address predicted2 = fixture.factory.predictEscrowAddress(
-            MERCHANT_ID_1,
-            MERCHANT_1,
-            address(fixture.daiToken)
-        );
+        address predicted2 = fixture.factory.predictEscrowAddress(MERCHANT_ID_1, MERCHANT_1, address(fixture.daiToken));
 
         assertTrue(predicted1 != predicted2);
     }
@@ -227,31 +161,19 @@ contract CloakEscrowFactoryDeploymentTest is Test, CloakEscrowFactoryFixtures {
     function test_DeployEscrow_RevertNotOwner() public {
         vm.prank(USER);
         vm.expectRevert();
-        fixture.factory.deployEscrow(
-            MERCHANT_ID_1,
-            MERCHANT_1,
-            address(fixture.usdcToken)
-        );
+        fixture.factory.deployEscrow(MERCHANT_ID_1, MERCHANT_1, address(fixture.usdcToken));
     }
 
     function test_DeployEscrow_RevertZeroMerchantAddress() public {
         // No prank needed - test contract is owner
         vm.expectRevert(CloakEscrowFactory.InvalidAddress.selector);
-        fixture.factory.deployEscrow(
-            MERCHANT_ID_1,
-            address(0),
-            address(fixture.usdcToken)
-        );
+        fixture.factory.deployEscrow(MERCHANT_ID_1, address(0), address(fixture.usdcToken));
     }
 
     function test_DeployEscrow_RevertZeroTokenAddress() public {
         // No prank needed - test contract is owner
         vm.expectRevert(CloakEscrowFactory.InvalidAddress.selector);
-        fixture.factory.deployEscrow(
-            MERCHANT_ID_1,
-            MERCHANT_1,
-            address(0)
-        );
+        fixture.factory.deployEscrow(MERCHANT_ID_1, MERCHANT_1, address(0));
     }
 
     function test_DeployEscrow_RevertWhenPaused() public {
@@ -260,24 +182,16 @@ contract CloakEscrowFactoryDeploymentTest is Test, CloakEscrowFactoryFixtures {
 
         // No prank needed - test contract is owner
         vm.expectRevert();
-        fixture.factory.deployEscrow(
-            MERCHANT_ID_1,
-            MERCHANT_1,
-            address(fixture.usdcToken)
-        );
+        fixture.factory.deployEscrow(MERCHANT_ID_1, MERCHANT_1, address(fixture.usdcToken));
     }
 
     // ============ Gas Optimization Tests ============
 
     function test_DeployEscrow_GasUsage() public {
         // No prank needed - test contract is owner
-        
+
         uint256 gasBefore = gasleft();
-        fixture.factory.deployEscrow(
-            MERCHANT_ID_1,
-            MERCHANT_1,
-            address(fixture.usdcToken)
-        );
+        fixture.factory.deployEscrow(MERCHANT_ID_1, MERCHANT_1, address(fixture.usdcToken));
         uint256 gasUsed = gasBefore - gasleft();
 
         // Gas usage should be reasonable (adjust threshold as needed)
